Update Example_Interface_3.ts
The supporting text for this example reads as follows:
"In the preceding snippet, we build a product object and then pass it to our class's **makeProduct** function. We then console out the results, which is the same as before, except now our functional code is wrapped in a class."

So I'm assuming line 34 should be `newProduct.makeProduct()` instead of `newProduct.product`.